data-bot/requirements.txt
CacheControl==0.12.10
Flask==2.0.3
Jinja2==3.0.3
MarkupSafe==2.1.0
Werkzeug==2.0.3
beautifulsoup4==4.10.0
bs4==0.0.1
cachetools==5.0.0
certifi==2021.10.8
certifi==2021.10.8
charset-normalizer==2.0.12
charset-normalizer==2.0.12
click==8.0.4
firebase-admin==5.2.0
google-api-core==2.7.1
google-api-python-client==2.40.0
googleapis-common-protos==1.55.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-cloud-core==2.2.3
google-cloud-firestore==2.4.0
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
grpcio==1.44.0
grpcio-status==1.44.0
httplib2==0.20.4
idna==3.3
idna==3.3
itsdangerous==2.1.1
lxml==4.8.0
msgpack==1.0.3
numpy==1.21.5
pandas==1.3.5
protobuf==3.19.4
proto-plus==1.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
requests==2.27.1
rsa==4.8
six==1.16.0
six==1.16.0
soupsieve==2.3.1
uritemplate==4.1.1
urllib3==1.26.8
urllib3==1.26.8
